refactor[#554] : useOverlay 제거하고 useModal 커스텀 훅 하나로 로직 통합
